simplified getBuildSpecificAsset() to something that definitely excludes assets
